process: fix unix behavior
