--- src/dummy.cpp
@@ -1,7 +1,0 @@
-/**
- * Mozart++ Template Library
- * Licensed under MIT License
- * Copyright (c) 2020 Covariant Institute
- * Website: https://covariant.cn/
- * Github:  https://github.com/covariant-institute/
- */
